[dnssec] fix compilation warnings on NetBSD

+++ src/dnssec/lib/keyid.c
@@ -38,7 +38,7 @@ bool dnssec_keyid_is_valid(const char *id)
 	}
 
 	for (int i = 0; i < DNSSEC_KEYID_SIZE; i++) {
-		if (!isxdigit(id[i])) {
+		if (!isxdigit((int)id[i])) {
 			return false;
 		}
 	}
@@ -54,8 +54,8 @@ void dnssec_keyid_normalize(char *id)
 	}
 
 	for (size_t i = 0; i < DNSSEC_KEYID_SIZE; i++) {
-		assert(id[i] != '\0' && isxdigit(id[i]));
-		id[i] = tolower(id[i]);
+		assert(id[i] != '\0' && isxdigit((int)id[i]));
+		id[i] = tolower((int)id[i]);
 	}
 }
